test(challenge): move challenge tests into challenge_test.go
Consolidate the whole-file/all-level traversal tests into the existing challenge_test.go instead of a separate file.
